♻️ converge on dispatch or defer for atoms and selectors
